# packages/records-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Records native API",
    "version": "1.0.0",
    "description": "Organisation datastores (Projects module). A request carries either a Records credential (`rk1_…`) together with a Cloudflare Access assertion for the API's own application, or a JWT from a trusted issuer (a delegated token or an Access for SaaS token), each bound to one datastore; the `datastoreId` in the path is a selector and must equal the credential's. Mutations need `Idempotency-Key`; changes to an existing record need `If-Match` with the `ETag` last read. Errors are RFC 9457 problem documents with a stable `code`."
  },
  "servers": [
    {
      "url": "/",
      "description": "Records service origin"
    }
  ],
...
  "security": [
    {
      "accessAssertion": [],
      "recordsCredential": []
    },
    {
      "delegatedToken": []
    }
  ],
  "paths": {
...
    "/gatekeeper/records/v1/datastores/{datastoreId}/poke": {
      "get": {
        "operationId": "subscribePokes",
        "summary": "Subscribe to pokes (`{datastoreId, head}` after each commit)",
        "description": "On Workers: a WebSocket upgrade (`Upgrade: websocket`) served by the datastore's Durable Object poke hub. On Node (@records/node): Server-Sent Events (`Accept: text/event-stream`), `event: poke` with the same JSON. Pokes carry no data and are at-most-once; subscribers pull on every poke and on a timer. The server closes a subscription after a maximum lifetime, so a revoked credential stops receiving pokes; reconnect and pull.",
        "tags": [
          "realtime"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/datastoreId"
          }
        ],
        "responses": {
          "101": {
            "description": "Switching protocols (WebSocket, on Workers)"
          },
          "200": {
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "x-records-idempotency": "safe",
        "x-records-sdk": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string",
            "enum": [
              "unauthenticated",
              "forbidden",
              "not_found",
              "validation_failed",
              "revision_conflict",
              "revision_required",
              "workflow_conflict",
              "idempotency_conflict",
              "datastore_archived",
              "duplicate",
              "payload_too_large",
              "rate_limited",
              "unavailable",
              "internal"
            ]
          },
          "detail": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ]
      },
...
    },
    "securitySchemes": {
      "accessAssertion": {
        "type": "apiKey",
        "in": "header",
        "name": "Cf-Access-Jwt-Assertion",
        "description": "Cloudflare Access assertion for the API's path-specific application, verified by the service (issuer + audience). Scripts usually obtain it by presenting an Access service token (`CF-Access-Client-Id` / `CF-Access-Client-Secret`) at the edge. A Node deployment may configure another front-door check or none."
      },
      "recordsCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "rk1",
        "description": "A datastore credential `rk1_<32 hex>_<43 base64url>`: a service principal with fixed scopes on one datastore."
      },
      "delegatedToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A JWT from a trusted issuer (records.trusted_issuers): a delegated token minted by the cloudflare-os Records gatekeeper after it redeems a viewer assertion (60 s, `sub` = the viewer's principal, `act` = the binding, `scope` = the binding's scopes, `aud` = the datastore service), or an Access for SaaS token for a mapped person."
      }
    },
    "parameters": {
      "datastoreId": {
        "name": "datastoreId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Must equal the credential's datastore."
      },
...
    },
    "headers": {
...
      "Retry-After": {
        "description": "Seconds to wait before retrying.",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      },
...
    },
    "responses": {
...
      "Unauthenticated": {
        "description": "`unauthenticated`: missing or invalid Access assertion or Records credential.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "`forbidden`: the credential's scopes or the principal's role do not allow this.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "`not_found`: no such route, record, or a datastore other than the credential's.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "`rate_limited`: back off for Retry-After seconds.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        }
      },
      "InternalError": {
        "description": "`internal`: an unexpected failure (no internals in the body).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unavailable": {
        "description": "`unavailable`: the request exceeded its 15 s deadline or the database was busy. Idempotent calls may be retried.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  },
...
}