# packages/records-sdk/src/generated/types.ts
// Generated by packages/records-sdk/scripts/generate.ts from the Records OpenAPI document. Do not edit.

export type PrincipalRef = {
  id: string;
  displayName: string;
  kind: "human" | "service";
};

export type Datastore = {
  id: string;
  name: string;
  description: string;
  moduleId: string;
  apiMajor: number;
  features: string[];
  lifecycle: "active" | "archived";
  ownerTeam: string | null;
  discovery: "members" | "organisation";
  role: "owner" | "admin" | "editor" | "reader" | null;
  createdAt: string;
  updatedAt: string;
  owner: PrincipalRef;
  retentionPolicy: string;
  environment: string;
  placement: string;
  moduleVersion: string | null;
  memberCount: number;
  activeBindingCount: number;
  activeCredentialCount: number;
  revision: number;
} | {
  id: string;
  name: string;
  description: string;
  moduleId: string;
  apiMajor: number;
  features: string[];
  lifecycle: "active" | "archived";
  ownerTeam: string | null;
  discovery: "members" | "organisation";
  role: "owner" | "admin" | "editor" | "reader" | null;
  createdAt: string;
  updatedAt: string;
};

export type Project = {
  id: string;
  key: string;
  name: string;
  description: string;
  revision: number;
  createdAt: string;
  updatedAt: string;
};

export type ProjectList = {
  items: Project[];
};

export type Workflow = {
  states: Array<{
    key: string;
    name: string;
    category: "todo" | "in_progress" | "done";
    position: number;
  }>;
  transitions: Array<{
    from: string;
    to: string;
  }>;
};

export type Issue = {
  id: string;
  projectId: string;
  number: number;
  key: string;
  title: string;
  description: string;
  state: string;
  priority: "none" | "low" | "medium" | "high" | "urgent";
  assignee: PrincipalRef | null;
  customFields: Record<string, string | number | boolean | null>;
  revision: number;
  createdAt: string;
  updatedAt: string;
  createdBy: PrincipalRef;
  updatedBy: PrincipalRef;
};

export type IssuePage = {
  items: Issue[];
  nextCursor: string | null;
};

export type Comment = {
  id: string;
  issueId: string;
  body: string;
  author: PrincipalRef;
  createdAt: string;
};

export type CommentPage = {
  items: Comment[];
  nextCursor: string | null;
};

export type AuditEvent = {
  id: string;
  datastoreId: string | null;
  operation: string;
  actor: PrincipalRef;
  initiator: PrincipalRef | null;
  bindingId: string | null;
  via: "gadget" | "http" | "management" | "system";
  targetType: string | null;
  targetId: string | null;
  summary: string;
  at: string;
};

export type AuditPage = {
  items: AuditEvent[];
  nextCursor: string | null;
};

export type JournalEntry = {
  seq: number;
  ordinal: number;
  changeId: string;
  command: string;
  entityType: "project" | "issue" | "comment";
  entityId: string;
  entityRev: number;
  op: "create" | "update" | "archive" | "restore" | "redact";
  after: Record<string, unknown>;
  before: Record<string, unknown> | null;
  actorId: string;
  actId: string | null;
  via: "gadget" | "http" | "sync" | "jira" | "management" | "system";
  occurredAt: string;
};

export type History = {
  items: JournalEntry[];
};

export type ChangesPage = {
  entries: JournalEntry[];
  nextAfter: number;
  head: number;
  resetRequired: boolean;
};

export type PushOutcome = {
  id: number;
  status: "applied";
  seq: number;
} | {
  id: number;
  status: "pending";
  actionId: number;
} | {
  id: number;
  status: "rejected";
  code: string;
  message: string;
} | {
  id: number;
  status: "conflict";
  code: "revision_conflict" | "workflow_conflict";
  message: string;
  currentRevision?: number;
} | {
  id: number;
  status: "skipped";
};

export type PushResponse = {
  outcomes: PushOutcome[];
  head: number;
};

export type PatchOp = {
  op: "clear";
} | {
  op: "put";
  key: string;
  value: unknown;
} | {
  op: "del";
  key: string;
};

export type PullResponse = {
  cookie: number;
  lastMutationIdChanges: Record<string, number>;
  patch: PatchOp[];
};

export type Problem = {
  type: string;
  title: string;
  status: number;
  code: "unauthenticated" | "forbidden" | "not_found" | "validation_failed" | "revision_conflict" | "revision_required" | "workflow_conflict" | "idempotency_conflict" | "datastore_archived" | "duplicate" | "payload_too_large" | "rate_limited" | "unavailable" | "internal";
  detail?: string;
  issues?: Array<{
    path: string;
    message: string;
  }>;
};

export type CreateIssueRequest = {
  id?: string;
  projectId: string;
  title: string;
  description?: string;
  state?: string;
  priority?: "none" | "low" | "medium" | "high" | "urgent";
  assigneeId?: string | null;
  customFields?: Record<string, string | number | boolean | null>;
};

export type EditIssueRequest = {
  title?: string;
  description?: string;
  priority?: "none" | "low" | "medium" | "high" | "urgent";
  assigneeId?: string | null;
  customFields?: Record<string, string | number | boolean | null>;
};

export type TransitionIssueRequest = {
  toState: string;
};

export type AddCommentRequest = {
  body: string;
};

export type PushRequest = {
  clientGroupId: string;
  clientId: string;
  mutations: Array<{
    id: number;
    name: "projects.createIssue" | "projects.editIssue" | "projects.transitionIssue" | "projects.addComment";
    args: unknown;
    timestamp?: number;
  }>;
};

export type PullRequest = {
  clientGroupId: string;
  cookie: number | null;
};

/** Query parameters of `listIssues`. */
export type ListIssuesQuery = {
  projectId?: string;
  state?: string;
  assigneeId?: string;
  /** Full-text query over title and description. */
  q?: string;
  order?: "updated_desc" | "number_asc";
  /** Page size (default 50, at most 100). */
  limit?: number;
  /** The previous page's `nextCursor`; valid only for the same query. */
  cursor?: string;
};

/** Query parameters of `listComments`. */
export type ListCommentsQuery = {
  /** Page size (default 50, at most 100). */
  limit?: number;
  /** The previous page's `nextCursor`; valid only for the same query. */
  cursor?: string;
};

/** Query parameters of `listChanges`. */
export type ListChangesQuery = {
  /** The last `seq` already seen (0 for the start). */
  after?: number;
  limit?: number;
};

/** Query parameters of `listAudit`. */
export type ListAuditQuery = {
  /** Page size (default 50, at most 100). */
  limit?: number;
  /** The previous page's `nextCursor`; valid only for the same query. */
  cursor?: string;
};